move more stuff to portfolio
