+++ src/containers/StartupProjects/StartupProject.js
@@ -1,60 +1,12 @@
 import React, { useContext, useState } from "react";
 import { Fade } from "react-reveal";
-import StyleContext from "../../contexts/StyleContext";
-
 import { Modal } from "../../components/modal/modal";
-import FeryaEventDecor from "../ferya/FeryaEventDecor";
+import StyleContext from "../../contexts/StyleContext";
+import { usePortfolio } from "../../portfolio";
 import "./StartupProjects.scss";
-import TheLink from "../theLink/TheLink";
-
-const bigProjects = {
-  title: "Projects",
-  subtitle: "INNOVATIVE INITIATIVES I'VE SHAPED FROM CONCEPT TO EXECUTION",
-  projects: [
-    {
-      image: require("../../assets/images/ferya.jpg"),
-      projectName: "Ferya Event Decor & Design",
-      role: "Co-Founder & Product Manager",
-      date: "September 2024 – Ongoing",
-      projectDesc: "Co-founded and developed a unique event decoration brand, defining its identity, business strategy, and client-focused solutions. Managed the product side, including service design, pricing strategies, and customer experience optimization.",
-      footerLink: [
-        {
-          isHyperlink: true,
-          name: "Visit Instagram",
-          url: "https://www.instagram.com/ferya.eventdecor?igsh=YTRvdDlpY2hocXM5"
-        },
-        {
-          isHyperlink: false,
-          name: "More",
-          component: <FeryaEventDecor />
-        }
-        //  you can add extra buttons here.
-      ]
-    },
-    {
-      image: require("../../assets/images/theLink.jpeg"),
-      projectName: "The Link",
-      role: "Product Manager",
-      date: "February 2025 – Ongoing",
-      projectDesc: "Contributed to an innovative stealth project aimed at transforming entrepreneurship through a scalable, user-centric platform. Managed the product side, including feature prioritization, user research, and roadmap development to align with business goals.",
-      footerLink: [
-        {
-          isHyperlink: true,
-          name: "Visit Website",
-          url: "https://the-link.ca/"
-        },
-        {
-          isHyperlink: false,
-          name: "More",
-          component: <TheLink />
-        }
-      ]
-    }
-  ],
-  display: true // Set false to hide this section, defaults to true
-};
 
 export default function StartupProject() {
+  const { bigProjects } = usePortfolio();
   const [showModal, setshowModal] = useState(false);
   const [modalContent, setmodalContent] = useState();
